+++ Assets/Scripts/Videos/VideoMappingSettings.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
-using UnityEngine;
 using VoyagerApp.Lamps;
 using VoyagerApp.Utilities;